:art: Show workbooks by type  (#698)
